Rename as `prefer-single-call`
