+++ rules/prefer-single-call.js
