+++ test/prefer-single-call.js
