+++ test/snapshots/prefer-single-call.js.md
@@ -1,6 +1,6 @@
-# Snapshot report for `test/prefer-combinable-action.js`
+# Snapshot report for `test/prefer-single-call.js`
 
-The actual snapshot is saved in `prefer-combinable-action.js.snap`.
+The actual snapshot is saved in `prefer-single-call.js.snap`.
 
 Generated by [AVA](https://avajs.dev).
